magento/inventory#3169: In store pickup intersection search graphql. Apply code review changes.
