+++ InventoryInStorePickupGraphQl/Model/Resolver/PickupLocations/SearchRequest/PageSize.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest;
+
+use Magento\InventoryInStorePickupApi\Model\SearchRequestBuilderInterface;
+
+/**
+ * Resolve Page Size and parameters.
+ */
+class PageSize implements ResolverInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public function resolve(
+        SearchRequestBuilderInterface $searchRequestBuilder,
+        string $fieldName,
+        string $argumentName,
+        array $argument
+    ): SearchRequestBuilderInterface {
+        $searchRequestBuilder->setPageSize($argument[$argumentName]);
+
+        return $searchRequestBuilder;
+    }
+}
