+++ InventoryInStorePickupGraphQl/Model/Resolver/PickupLocations/SearchRequestBuilder.php
@@ -7,14 +7,14 @@ declare(strict_types=1);
 
 namespace Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations;
 
-use Magento\Framework\Api\SortOrder;
-use Magento\Framework\Api\SortOrderBuilder;
-use Magento\Framework\GraphQl\Query\Resolver\Argument\AstConverter;
-use Magento\Framework\GraphQl\Query\Resolver\Argument\Filter\Clause;
-use Magento\InventoryApi\Api\Data\SourceInterface;
-use Magento\InventoryInStorePickupApi\Api\Data\PickupLocationInterface;
 use Magento\InventoryInStorePickupApi\Model\SearchRequestBuilderInterface;
 use Magento\InventoryInStorePickupApi\Model\SearchRequestBuilderInterfaceFactory;
+use Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\Area;
+use Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\CurrentPage;
+use Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\Filter;
+use Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\PageSize;
+use Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\ResolverInterface;
+use Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequest\Sort;
 use Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchRequestBuilder\ExtensionProvider;
 
 /**
@@ -22,23 +22,11 @@ use Magento\InventoryInStorePickupGraphQl\Model\Resolver\PickupLocations\SearchR
  */
 class SearchRequestBuilder
 {
-    private const RADIUS_FIELD = 'radius';
-    private const SEARCH_TERM = 'search_term';
     private const SORT_FIELD = 'sort';
     private const AREA_FIELD = 'area';
     private const FILTERS_FIELD = 'filters';
     private const CURRENT_PAGE = 'currentPage';
     private const PAGE_SIZE = 'pageSize';
-
-    /**
-     * @var AstConverter
-     */
-    private $astConverter;
-
-    /**
-     * @var SortOrderBuilder
-     */
-    private $sortOrderBuilder;
 
     /**
      * @var ExtensionProvider
@@ -51,21 +39,38 @@ class SearchRequestBuilder
     private $searchRequestBuilderFactory;
 
     /**
-     * @param AstConverter $astConverter
-     * @param SortOrderBuilder $sortOrderBuilder
+     * @var ResolverInterface[]
+     */
+    private $resolvers;
+
+    /**
      * @param ExtensionProvider $extensionProvider
      * @param SearchRequestBuilderInterfaceFactory $searchRequestBuilderFactory
+     * @param Area $area
+     * @param CurrentPage $currentPage
+     * @param Filter $filter
+     * @param PageSize $pageSize
+     * @param Sort $sort
      */
     public function __construct(
-        AstConverter $astConverter,
-        SortOrderBuilder $sortOrderBuilder,
         ExtensionProvider $extensionProvider,
-        SearchRequestBuilderInterfaceFactory $searchRequestBuilderFactory
+        SearchRequestBuilderInterfaceFactory $searchRequestBuilderFactory,
+        Area $area,
+        CurrentPage $currentPage,
+        Filter $filter,
+        PageSize $pageSize,
+        Sort $sort
     ) {
-        $this->astConverter = $astConverter;
-        $this->sortOrderBuilder = $sortOrderBuilder;
         $this->extensionProvider = $extensionProvider;
         $this->searchRequestBuilderFactory = $searchRequestBuilderFactory;
+
+        $this->resolvers = [
+            self::AREA_FIELD => $area,
+            self:: CURRENT_PAGE => $currentPage,
+            self::PAGE_SIZE => $pageSize,
+            self::SORT_FIELD => $sort,
+            self::FILTERS_FIELD => $filter
+        ];
     }
 
     /**
@@ -76,182 +81,18 @@ class SearchRequestBuilder
      *
      * @return SearchRequestBuilderInterface
      */
-    public function getBuilderFromArgument(string $fieldName, array $argument): SearchRequestBuilderInterface
+    public function getFromArgument(string $fieldName, array $argument): SearchRequestBuilderInterface
     {
         $searchRequestBuilder = $this->searchRequestBuilderFactory->create();
 
         foreach (array_keys($argument) as $argumentName) {
-            $this->resolveAttributeValue($searchRequestBuilder, $fieldName, $argumentName, $argument);
+            if (isset($this->resolvers[$argumentName])) {
+                $this->resolvers[$argumentName]->resolve($searchRequestBuilder, $fieldName, $argumentName, $argument);
+            }
         }
 
         $searchRequestBuilder->setSearchRequestExtension($this->extensionProvider->getExtensionAttributes($argument));
 
         return $searchRequestBuilder;
-    }
-
-    /**
-     * Resolve input attributes values and pass them to Search Request Builder.
-     *
-     * @param SearchRequestBuilderInterface $searchRequestBuilder
-     * @param string $fieldName
-     * @param string $argumentName
-     * @param array $argument
-     *
-     * @return void
-     */
-    private function resolveAttributeValue(
-        SearchRequestBuilderInterface $searchRequestBuilder,
-        string $fieldName,
-        string $argumentName,
-        array $argument
-    ): void {
-        switch ($argumentName) {
-            case self::AREA_FIELD:
-                $searchRequestBuilder->setAreaSearchTerm($this->getAreaSearchTerm($argument));
-                $searchRequestBuilder->setAreaRadius($this->getAreaRadius($argument));
-                break;
-            case self::FILTERS_FIELD:
-                $filters = $this->getFilters($fieldName, $argument);
-                foreach ($filters as $filter) {
-                    $this->addFilterToBuilder($searchRequestBuilder, $filter);
-                }
-                break;
-            case self::CURRENT_PAGE:
-                $searchRequestBuilder->setCurrentPage($argument[$argumentName]);
-                break;
-            case self::PAGE_SIZE:
-                $searchRequestBuilder->setPageSize($argument[$argumentName]);
-                break;
-            case self::SORT_FIELD:
-                $searchRequestBuilder->setSortOrders($this->getSortOrders($argument));
-                break;
-        }
-    }
-
-    /**
-     * Get Radius attribute value.
-     *
-     * @param array $argument
-     *
-     * @return int
-     */
-    private function getAreaRadius(array $argument): int
-    {
-        return $argument[self::AREA_FIELD][self::RADIUS_FIELD];
-    }
-
-    /**
-     * Get Search Term attribute value.
-     *
-     * @param array $argument
-     *
-     * @return string
-     */
-    private function getAreaSearchTerm(array $argument): string
-    {
-        return $argument[self::AREA_FIELD][self::SEARCH_TERM];
-    }
-
-    /**
-     * Get filters from arguments.
-     *
-     * @param string $fieldName
-     * @param array $argument
-     *
-     * @return array
-     */
-    private function getFilters(string $fieldName, array $argument): array
-    {
-        return $this->astConverter->getClausesFromAst($fieldName, $argument[self::FILTERS_FIELD]);
-    }
-
-    /**
-     * Get Sort Orders from arguments.
-     *
-     * @param array $argument
-     *
-     * @return array
-     */
-    private function getSortOrders(array $argument): array
-    {
-        $sortOrders = [];
-        foreach ($argument[self::SORT_FIELD] as $fieldName => $fieldValue) {
-            /** @var SortOrder $sortOrder */
-            $sortOrders[] = $this->sortOrderBuilder
-                ->setField($fieldName)
-                ->setDirection(
-                    $fieldValue === SortOrder::SORT_DESC ? SortOrder::SORT_DESC : SortOrder::SORT_ASC
-                )->create();
-        }
-
-        return $sortOrders;
-    }
-
-    /**
-     * Add filter to the Search Request Builder.
-     *
-     * @param SearchRequestBuilderInterface $searchRequestBuilder
-     * @param Clause $filter
-     */
-    private function addFilterToBuilder(SearchRequestBuilderInterface $searchRequestBuilder, Clause $filter): void
-    {
-        switch ($filter->getFieldName()) {
-            case SourceInterface::COUNTRY_ID:
-                $searchRequestBuilder->setCountryFilter(
-                    $this->getClauseValue($filter),
-                    $filter->getClauseType()
-                );
-                break;
-            case SourceInterface::POSTCODE:
-                $searchRequestBuilder->setPostcodeFilter(
-                    $this->getClauseValue($filter),
-                    $filter->getClauseType()
-                );
-                break;
-            case SourceInterface::REGION:
-                $searchRequestBuilder->setRegionFilter(
-                    $this->getClauseValue($filter),
-                    $filter->getClauseType()
-                );
-                break;
-            case SourceInterface::REGION_ID:
-                $searchRequestBuilder->setRegionIdFilter(
-                    $this->getClauseValue($filter),
-                    $filter->getClauseType()
-                );
-                break;
-            case SourceInterface::CITY:
-                $searchRequestBuilder->setCityFilter($this->getClauseValue($filter), $filter->getClauseType());
-                break;
-            case SourceInterface::STREET:
-                $searchRequestBuilder->setStreetFilter(
-                    $this->getClauseValue($filter),
-                    $filter->getClauseType()
-                );
-                break;
-            case PickupLocationInterface::PICKUP_LOCATION_CODE:
-                $searchRequestBuilder->setPickupLocationCodeFilter(
-                    $this->getClauseValue($filter),
-                    $filter->getClauseType()
-                );
-                break;
-            case SourceInterface::NAME:
-                $searchRequestBuilder->setNameFilter($this->getClauseValue($filter), $filter->getClauseType());
-                break;
-        }
-    }
-
-    /**
-     * Get value from the clause.
-     *
-     * @param Clause $filter
-     *
-     * @return string
-     */
-    private function getClauseValue(Clause $filter): string
-    {
-        $value = $filter->getClauseValue();
-
-        return is_array($value) ? implode(',', $value) : $value;
     }
 }
